requirements.txt
typer
rich
python-crontab
psycopg2-binary
mysql-connector-python
python-dotenv
pymssql
mssql-scripter
InquirerPy
